requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
python-dotenv==1.0.1
langchain-core>=0.3.36
langchain-anthropic>=0.1.4
langgraph>=0.0.26
langchain-mcp-adapters>=0.0.1
pydantic>=2.6.1
python-multipart==0.0.9 